add ipad scale, base label button

+++ trunk/Vocabulary/Game/Classes/General/GameManager.m
@@ -38,7 +38,7 @@
 }
 
 -(Winner)calculateWinner {
-    NSInteger winner = Tie;
+    int winner = Tie;
     if (self.playerOneScore > self.playerTwoScore) {
         winner = PlayerOneWin;
     } else if (self.playerOneScore < self.playerTwoScore) {
